+++ Plugins/MarkdownAsset/Source/MarkdownAssetEditor/Private/MarkitdownSettings.cpp
@@ -1,0 +1,13 @@
+// Copyright (c) 2026 Kurorekishi (EmbarrassingMoment).
+
+#include "MarkitdownSettings.h"
+
+UMarkitdownSettings::UMarkitdownSettings()
+	: ExecutionMode(EMarkitdownExecutionMode::Uvx)
+{
+}
+
+FName UMarkitdownSettings::GetCategoryName() const
+{
+	return TEXT("Plugins");
+}